decoupled AddressLogic by creating an interface that handles all Google Api requests
