+++ address/src/Revature.Address.Lib/BusinessLogic/IGoogleApiAccess.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Revature.Address.Lib.BusinessLogic
+{
+  public interface IGoogleApiAccess
+  {
+    public Task<double> GetDistance(Address origin, Address destination, int distance);
+    public Task<bool> IsValidAddressAsync(Address address);
+    public Task<Address> NormalizeAddressAsync(Address address);
+  }
+}
